Migrando Scoreboard a frontend/src
Este cambio mueve \OmegaUp\Scoreboard al directorio que puede ser
autoloaded.
