+++ frontend/server/controllers/ContestController.php
@@ -1348,7 +1348,7 @@ class ContestController extends Controller {
 
         // Invalidar cache
         \OmegaUp\Cache::deleteFromCache(\OmegaUp\Cache::CONTEST_INFO, $r['contest_alias']);
-        Scoreboard::invalidateScoreboardCache(ScoreboardParams::fromContest($params['contest']));
+        \OmegaUp\Scoreboard::invalidateScoreboardCache(\OmegaUp\ScoreboardParams::fromContest($params['contest']));
 
         return ['status' => 'ok'];
     }
@@ -1434,7 +1434,7 @@ class ContestController extends Controller {
 
         // Invalidar cache
         \OmegaUp\Cache::deleteFromCache(\OmegaUp\Cache::CONTEST_INFO, $r['contest_alias']);
-        Scoreboard::invalidateScoreboardCache(ScoreboardParams::fromContest($params['contest']));
+        \OmegaUp\Scoreboard::invalidateScoreboardCache(\OmegaUp\ScoreboardParams::fromContest($params['contest']));
 
         return ['status' => 'ok'];
     }
@@ -1797,13 +1797,13 @@ class ContestController extends Controller {
         // Get the current user
         $response = self::validateDetails($r);
 
-        $params = ScoreboardParams::fromContest($response['contest']);
-        $params['admin'] = (
+        $params = \OmegaUp\ScoreboardParams::fromContest($response['contest']);
+        $params->admin = (
             \OmegaUp\Authorization::isContestAdmin($r->identity, $response['contest']) &&
             !ContestsDAO::isVirtual($response['contest'])
         );
-        $params['show_all_runs'] = !ContestsDAO::isVirtual($response['contest']);
-        $scoreboard = new Scoreboard($params);
+        $params->show_all_runs = !ContestsDAO::isVirtual($response['contest']);
+        $scoreboard = new \OmegaUp\Scoreboard($params);
 
         // Push scoreboard data in response
         $response = [];
@@ -1845,9 +1845,9 @@ class ContestController extends Controller {
         }
 
         // Create scoreboard
-        $params = ScoreboardParams::fromContest($contest);
-        $params['admin'] = $showAllRuns;
-        $scoreboard = new Scoreboard($params);
+        $params = \OmegaUp\ScoreboardParams::fromContest($contest);
+        $params->admin = $showAllRuns;
+        $scoreboard = new \OmegaUp\Scoreboard($params);
 
         return $scoreboard->generate();
     }
@@ -1901,9 +1901,9 @@ class ContestController extends Controller {
                 $r['contest_params'] = $cp;
             }
 
-            $params = ScoreboardParams::fromContest($contest);
-            $params['only_ac'] = $r['contest_params'][$contest->alias]['only_ac'];
-            $s = new Scoreboard($params);
+            $params = \OmegaUp\ScoreboardParams::fromContest($contest);
+            $params->only_ac = $r['contest_params'][$contest->alias]['only_ac'];
+            $s = new \OmegaUp\Scoreboard($params);
 
             $scoreboards[$contest->alias] = $s->generate();
         }
@@ -2223,7 +2223,7 @@ class ContestController extends Controller {
         \OmegaUp\Cache::deleteFromCache(\OmegaUp\Cache::CONTEST_INFO, $r['contest_alias']);
 
         // Expire contest scoreboard cache
-        Scoreboard::invalidateScoreboardCache(ScoreboardParams::fromContest($contest));
+        \OmegaUp\Scoreboard::invalidateScoreboardCache(\OmegaUp\ScoreboardParams::fromContest($contest));
 
         // Expire contest-list cache
         \OmegaUp\Cache::invalidateAllKeys(\OmegaUp\Cache::CONTESTS_LIST_PUBLIC);
@@ -2508,10 +2508,10 @@ class ContestController extends Controller {
         self::authenticateRequest($r);
         $contest = self::validateStats($r['contest_alias'], $r->identity);
 
-        $params = ScoreboardParams::fromContest($contest);
-        $params['admin'] = true;
-        $params['auth_token'] = $r['auth_token'];
-        $scoreboard = new Scoreboard($params);
+        $params = \OmegaUp\ScoreboardParams::fromContest($contest);
+        $params->admin = true;
+        $params->auth_token = $r['auth_token'];
+        $scoreboard = new \OmegaUp\Scoreboard($params);
 
         // Check the filter if we have one
         \OmegaUp\Validators::validateStringNonEmpty($r['filterBy'], 'filterBy', false /* not required */);
